API: CreditCard: verification
